requirements.txt
python-pptx==0.6.19
google-cloud-vision==1.0.0
Pillow==7.2.0

